requirements.txt
asgiref
black
certifi
charset-normalizer
click
coreapi
coreschema
Django
djangorestframework
djangorestframework-simplejwt
drf-yasg
Faker
flake8
idna
inflection
itypes
Jinja2
MarkupSafe
mccabe
mypy-extensions
packaging
pathspec
pillow
platformdirs
pycodestyle
pyflakes
PyJWT
python-dateutil
pytz
PyYAML
requests
setuptools
six
sqlparse
uritemplate
urllib3